+++ WrightTools/artists/_interact.py
@@ -17,8 +17,9 @@ __all__ = ["interact2D"]
 
 
 class Focus:
-    def __init__(self, axes, linewidth=2):
+    def __init__(self, axes, sliders, linewidth=2):
         self.axes = axes
+        self.sliders = sliders
         self.linewidth = linewidth
         ax = axes[0]
         for side in ["top", "bottom", "left", "right"]:
@@ -36,6 +37,11 @@ class Focus:
         if self.focus_axis == ax or ax not in self.axes:
             return
         else:  # set new focus
+            if self.focus_axis.get_gid() in self.sliders.keys():
+                self.sliders[self.focus_axis.get_gid()].track.set_facecolor("lightgrey")
+            if ax.get_gid() in self.sliders.keys():
+                self.sliders[ax.get_gid()].track.set_facecolor("darkgrey")
+
             for spine in ["top", "bottom", "left", "right"]:
                 self.focus_axis.spines[spine].set_linewidth(1)
                 ax.spines[spine].set_linewidth(self.linewidth)
@@ -252,8 +258,10 @@ def interact2D(
             if axis.size > np.prod(axis.shape):
                 raise NotImplementedError("Cannot use multivariable axis as a slider")
             slider_axes = plt.subplot(gs[~len(sliders), :]).axes
-            slider = Slider(slider_axes, axis.label, 0, axis.points.size - 1, valinit=0, valstep=1)
+            slider = Slider(slider_axes, axis.label, 0, axis.points.size - 1, valinit=0, valstep=1, track_color="lightgrey")            
             sliders[axis.natural_name] = slider
+            slider_axes.set_gid(axis.natural_name)
+            # axis_to_slider[slider_axes.]
             slider.ax.vlines(
                 range(axis.points.size - 1),
                 *slider.ax.get_ylim(),
@@ -262,7 +270,7 @@ def interact2D(
                 alpha=0.5,
             )
             slider.valtext.set_text(gen_ticklabels(axis.points)[0])
-    current_state.focus = Focus([ax0] + [slider.ax for slider in sliders.values()])
+    current_state.focus = Focus([ax0] + [slider.ax for slider in sliders.values()], sliders)
     # initial xyz start are from zero indices of additional axes
     current_state.dat = data.chop(
         xaxis.natural_name,
@@ -375,6 +383,9 @@ def interact2D(
     if channel.signed:
         sp_x.set_ylim(-1.1, 1.1)
         sp_y.set_xlim(-1.1, 1.1)
+    else:
+        sp_x.set_ylim(0, 1.1)
+        sp_y.set_xlim(0, 1.1)
 
     def update_sideplot_slices():
         # TODO:  if bins is only available along one axis, slicing should be valid along the other
@@ -447,8 +458,16 @@ def interact2D(
         ticklabels = gen_ticklabels(ticks, channel.signed)
         colorbar.set_ticklabels(ticklabels)
 
-        sp_x.clear()
-        sp_y.clear()
+        [item.remove() for item in sp_x.collections]
+        [item.remove() for item in sp_y.collections]
+
+        if channel.signed:
+            sp_x.set_ylim(-1.1, 1.1)
+            sp_y.set_xlim(-1.1, 1.1)
+        else:
+            sp_x.set_ylim(0, 1.1)
+            sp_y.set_xlim(0, 1.1)
+
         draw_sideplot_projections()
         if line_sp_x.get_visible() and line_sp_y.get_visible():
             update_sideplot_slices()